requirements.txt
Flask==2.2.5
Flask-AppBuilder==4.3.7
Flask-APScheduler==1.13.1
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
PyJWT==2.9.0
pytest-flask==1.3.0
pre_commit==4.0.1
black==24.10.0
flake8==7.1.1
isort==5.13.2
